(fix)board, profile serializer
