+++ .idea/workspace.xml
@@ -2,8 +2,10 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="14847226-151c-4591-a157-0243461eac9c" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/Server/members/serializers.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/Server/boards/models.py" beforeDir="false" afterPath="$PROJECT_DIR$/Server/boards/models.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Server/boards/serializers.py" beforeDir="false" afterPath="$PROJECT_DIR$/Server/boards/serializers.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Server/profiles/serializers.py" beforeDir="false" afterPath="$PROJECT_DIR$/Server/profiles/serializers.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
